# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Emailed API
  description: |
    The Emailed API provides programmatic access to the Emailed email
    infrastructure platform. Send transactional and bulk emails, manage
    domains, configure webhooks, and query analytics.
  version: 0.1.0
  contact:
    name: Emailed Support
    url: https://emailed.dev/support
    email: support@emailed.dev
  license:
    name: Proprietary
    url: https://emailed.dev/terms
servers:
  - url: https://api.emailed.dev
    description: Production
  - url: https://api.staging.emailed.dev
    description: Staging
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /v1/voice/analyze:
    post:
      operationId: analyzeVoice
      summary: Build voice profile from sent emails
      tags: [Voice]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                sampleSize:
                  type: integer
                  default: 50
                  maximum: 200
      responses:
        "200":
          description: Voice profile built
        "400":
          description: Insufficient sent emails
  /v1/voice/profile:
    get:
      operationId: getVoiceProfile
      summary: Get current voice profile
      tags: [Voice]
      responses:
        "200":
          description: Voice profile data
        "404":
          description: No profile exists yet
  /v1/voice/draft:
    post:
      operationId: generateDraft
      summary: Generate email draft in user's voice
      tags: [Voice]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instructions]
              properties:
                instructions:
                  type: string
                tone:
                  type: string
                  enum: [professional, casual, friendly, formal, urgent, empathetic, assertive]
                  default: professional
                length:
                  type: string
                  enum: [brief, moderate, detailed]
                  default: moderate
                recipientName:
                  type: string
                subject:
                  type: string
                replyTo:
                  type: object
                  properties:
                    from:
                      type: string
                    subject:
                      type: string
                    body:
                      type: string
      responses:
        "200":
          description: Generated draft
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      subject:
                        type: string
                      body:
                        type: string
                      tone:
                        type: string
  /v1/voice/adjust:
    post:
      operationId: adjustTone
      summary: Adjust tone of existing email text
      tags: [Voice]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body, tone]
              properties:
                body:
                  type: string
                tone:
                  type: string
                  enum: [professional, casual, friendly, formal, urgent, empathetic, assertive]
      responses:
        "200":
          description: Tone-adjusted text
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Pass a Bearer token in the Authorization header.
        Example: `Authorization: Bearer em_live_abc123...`
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Pass an API key in the X-API-Key header.
        Example: `X-API-Key: em_live_abc123...`
...